pkg/advent_of_code/2021/07/day7_test.go
package y21d7

import (
  "testing"
  "github.com/fdm1/advent_of_code_go/pkg/aoc_utils"
)

func TestPart1(t *testing.T) {
  aoc_utils.MockAdventOfCodeInput(2021, 7)
  result := Part1(false)
  expectedResult := "37"

  if result != expectedResult {
    t.Fatalf("actual result %v != expected result %v", result, expectedResult)
  }
}

func TestPart2(t *testing.T) {
  aoc_utils.MockAdventOfCodeInput(2021, 7)
  result := Part2(false)
  expectedResult := "168"

  if result != expectedResult {
    t.Fatalf("actual result %v != expected result %v", result, expectedResult)
  }
}

func TestP2FuelCalculator(t *testing.T) {
  result := P2FuelCalculator(11)
  expectedResult := 66
  if result != expectedResult {
    t.Fatalf("actual result %v != expected result %v", result, expectedResult)
  }
}
